pyproject.toml
# File: pyproject.toml
[tool.poetry]
name = "kmeans_autoclust"
version = "0.1.0"
description = "Automated K-Means Clustering with multiple distance metrics"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^1.20.0"
pandas = "^1.5.0"
scikit-learn = "^1.2.0"
scipy = "^1.9.0"
tslearn = "^0.5.2"
gower = "^0.1.3"
pyemd = "^0.5.3"
python-Levenshtein = "^0.20.9"
sklearn-extra = "^0.4.0"
pyyaml = "^6.0"
python-json-logger = "^2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"